Added a show cards function and a deck for each player
I just added a basic show cards function that can be used at the end of rounds to give players 3 options, though its a bit wordy. One thing that has to be worked on for that though is the placement of the cards. I just placed the first card where our mouse is for now and the others a space to the right away. I also added a deck array for each player.